fix(main): change year and week types to int

+++ src/api/main.cpp
@@ -41,8 +41,8 @@ struct ApiOptions
 {
     std::string studyFolder = "";
     std::string outputFolder = "";
-    unsigned int year = -1;
-    unsigned int week = -1;
+    int year = -1;
+    int week = -1;
     bool writeMps = false;
 };
 
@@ -137,7 +137,8 @@ void printProblems(const ApiOptions& options)
         for (int week = firstWeek; week < lastWeek; ++week)
         {
             logs.info() << " week: " << week << '\n';
-            const WeeklyProblemId id = {year, week};
+            const WeeklyProblemId id = {static_cast<unsigned int>(year),
+                                        static_cast<unsigned int>(week)};
             auto weekly = getter.getWeeklyProblem(id);
             // printWeekLPData(constant, weekly);
             if (options.writeMps)
